+++ Configuration/TypoScript/GridElements/DocumentSectionGridElement/setup.ts
@@ -1,0 +1,25 @@
+# === Document Section Grid Element Content Records Configuration ===
+
+tt_content {
+  gridelements_pi1 {
+    20 {
+      10 {
+        setup {
+          t3vContentDocumentSection < lib.gridelements.defaultGridSetup
+          t3vContentDocumentSection {
+            cObject = FLUIDTEMPLATE
+            cObject {
+              file = EXT:t3v_content/Resources/Private/Templates/GridElements/DocumentSectionGridElement.html
+
+              layoutRootPath = EXT:t3v_content/Resources/Private/Layouts/GridElements/
+
+              partialRootPath = EXT:t3v_content/Resources/Private/Partials/GridElements/
+            }
+
+            outerWrap = |
+          }
+        }
+      }
+    }
+  }
+}