+++ Configuration/TypoScript/GridElements/setup.ts
@@ -2,4 +2,5 @@
 
 <INCLUDE_TYPOSCRIPT: source="FILE: ./AccordionGridElement/setup.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayoutGridElement/setup.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./DocumentSectionGridElement/setup.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./ViewportGridElement/setup.ts">